requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
dnspython==2.8.0
idna==3.11
IMAPClient==3.0.1
pymongo==4.15.5
python-dateutil==2.9.0.post0
pyzmail36==1.0.5
requests==2.32.5
six==1.17.0
urllib3==2.6.2
